put ISharedFileCache back into SharedFileCache, refactor out Browse()
